+++ src/modules/episodes/model/episodes.model.ts
@@ -1,6 +1,6 @@
 import {Prop, Schema, SchemaFactory} from '@nestjs/mongoose';
-import {DEFAULT_SCHEMA_OPTIONS} from "../../shared/const/schema.options";
-import {DefaultModel} from "../../shared/const/default.model";
+import {DEFAULT_SCHEMA_OPTIONS} from "../../../shared/const/schema.options";
+import {DefaultModel} from "../../../shared/const/default.model";
 import {EpisodeResponseObject} from "../dto/episodes.dto";
 import {Types} from "mongoose";
 import {Media, MediaSchema} from "./media.model";
