+++ src/modules/episodes/query/episodes.query.ts
@@ -1,4 +1,4 @@
-import {DEFAULT_QUERY_VALUES, IDefaultQuery} from "../../shared/const/default.query.interface";
+import {DEFAULT_QUERY_VALUES, IDefaultQuery} from "../../../shared/const/default.query.interface";
 
 export const DEFAULT_EPISODE_QUERY: IDefaultQuery = {
     ...DEFAULT_QUERY_VALUES,
