+++ src/modules/game-generation/game-generation.service.ts
@@ -4,8 +4,8 @@ import {InjectModel} from "@nestjs/mongoose";
 import {Model, Types} from "mongoose";
 import {Game} from "../games/model/games.model";
 import {GamesService} from "../games/games.service";
-import {ERROR_TYPES} from "../shared/const/error.types";
-import {asyncForEach} from "../shared/utils/utils";
+import {ERROR_TYPES} from "../../shared/const/error.types";
+import {asyncForEach} from "../../shared/utils/utils";
 import {Episode} from "../episodes/model/episodes.model";
 import {MediaConfig} from "../episodes/model/media.model";
 import {IgdbService} from "../igdb/igdb.service";
