+++ src/modules/games/games.controller.ts
@@ -1,11 +1,11 @@
 import {Body, Controller, Delete, Get, Logger, Param, Post, Put, Query, UseGuards, UsePipes} from '@nestjs/common';
 import {GamesService} from "./games.service";
 import {CreateGameDto, GameResponseObject, UpdateGameDto} from "./dto/games.dto";
-import {ValidationPipe} from "../shared/handler/validation.pipe";
+import {ValidationPipe} from "../../shared/handler/validation.pipe";
 import {DEFAULT_GAME_QUERY, GAME_SEARCHABLE_INDEX, IGameQuery} from "./query/games.query.interface";
-import {_parseDefaultQueryTypes} from "../shared/const/default.query.interface";
-import {removeEmptyAttrFromObj} from "../shared/utils/utils";
-import {AuthGuard} from "../shared/handler/auth.guard";
+import {_parseDefaultQueryTypes} from "../../shared/const/default.query.interface";
+import {removeEmptyAttrFromObj} from "../../shared/utils/utils";
+import {AuthGuard} from "../../shared/handler/auth.guard";
 
 @Controller('games')
 export class GamesController {
