+++ src/modules/games/games.service.ts
@@ -13,8 +13,8 @@ import {CreateGameDto, GameResponseObject, UpdateGameDto} from "./dto/games.dto"
 import {InjectModel} from "@nestjs/mongoose";
 
 import {Game} from "./model/games.model";
-import {asyncForEach, isObjectId} from "../shared/utils/utils";
-import {ERROR_TYPES} from "../shared/const/error.types";
+import {asyncForEach, isObjectId} from "../../shared/utils/utils";
+import {ERROR_TYPES} from "../../shared/const/error.types";
 import {Model, Types} from "mongoose";
 import {EpisodesService} from "../episodes/episodes.service";
 import {IGameQuery} from "./query/games.query.interface";
