+++ src/modules/games/query/games.query.interface.ts
@@ -1,4 +1,4 @@
-import {DEFAULT_QUERY_VALUES, IDefaultQuery} from "../../shared/const/default.query.interface";
+import {DEFAULT_QUERY_VALUES, IDefaultQuery} from "../../../shared/const/default.query.interface";
 
 export interface IGameQuery extends IDefaultQuery {
     name?: string,
