+++ src/tasks/tasks.module.ts
@@ -1,6 +1,6 @@
 import {forwardRef, Module} from '@nestjs/common';
 import {TasksService} from './tasks.service';
-import {EpisodesModule} from "../episodes/episodes.module";
+import {EpisodesModule} from "../modules/episodes/episodes.module";
 
 @Module({
     imports: [forwardRef(() => EpisodesModule)],
